requirements.txt
python-binance==1.0.35
pandas>=2.2.3
pandas-ta==0.4.71b0
numpy>=2.2.6
aiohttp>=3.10.0
python-dotenv==1.0.1
